portfolio-backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
drf-yasg>=1.21
markdown>=3.5
django-cors-headers>=4.3
python-dotenv>=1.0.0
